lib/models.rb  -->  models/*.rb に分割移動
